requirements.txt
aioconsole==0.8.0
aiohttp==3.10.10
asyncpg==0.29.0
SQLAlchemy==2.0.35
greenlet==3.1.1
python-dotenv==1.0.1
XlsxWriter==3.2.0